+++ dev-packages/browser-integration-tests/suites/tracing/trace-lifetime/pageload-meta/test.ts
@@ -1,7 +1,8 @@
 import { expect } from '@playwright/test';
-import type { Event } from '@sentry/types';
 import { sentryTest } from '../../../../utils/fixtures';
+import type { EventAndTraceHeader } from '../../../../utils/helpers';
 import {
+  eventAndTraceHeaderRequestParser,
   getFirstSentryEnvelopeRequest,
   getMultipleSentryEnvelopeRequests,
   shouldSkipTracingTest,
@@ -10,7 +11,7 @@ import {
 const META_TAG_TRACE_ID = '12345678901234567890123456789012';
 const META_TAG_PARENT_SPAN_ID = '1234567890123456';
 const META_TAG_BAGGAGE =
-  'sentry-trace_id=12345678901234567890123456789012,sentry-sample_rate=0.2,sentry-transaction=my-transaction,sentry-public_key=public,sentry-release=1.0.0,sentry-environment=prod';
+  'sentry-trace_id=12345678901234567890123456789012,sentry-sample_rate=0.2,sentry-sampled=true,sentry-transaction=my-transaction,sentry-public_key=public,sentry-release=1.0.0,sentry-environment=prod';
 
 sentryTest(
   'create a new trace for a navigation after the <meta> tag pageload trace',
@@ -21,8 +22,16 @@ sentryTest(
 
     const url = await getLocalTestPath({ testDir: __dirname });
 
-    const pageloadEvent = await getFirstSentryEnvelopeRequest<Event>(page, url);
-    const navigationEvent = await getFirstSentryEnvelopeRequest<Event>(page, `${url}#foo`);
+    const [pageloadEvent, pageloadTraceHeader] = await getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+      page,
+      url,
+      eventAndTraceHeaderRequestParser,
+    );
+    const [navigationEvent, navigationTraceHeader] = await getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+      page,
+      `${url}#foo`,
+      eventAndTraceHeaderRequestParser,
+    );
 
     const pageloadTraceContext = pageloadEvent.contexts?.trace;
     const navigationTraceContext = navigationEvent.contexts?.trace;
@@ -33,6 +42,17 @@ sentryTest(
       parent_span_id: META_TAG_PARENT_SPAN_ID,
       span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
     });
+
+    expect(pageloadTraceHeader).toEqual({
+      environment: 'prod',
+      release: '1.0.0',
+      sample_rate: '0.2',
+      sampled: 'true',
+      transaction: 'my-transaction',
+      public_key: 'public',
+      trace_id: META_TAG_TRACE_ID,
+    });
+
     expect(navigationTraceContext).toMatchObject({
       op: 'navigation',
       trace_id: expect.stringMatching(/^[0-9a-f]{32}$/),
@@ -40,6 +60,14 @@ sentryTest(
     });
     // navigation span is head of trace, so there's no parent span:
     expect(navigationTraceContext?.trace_id).not.toHaveProperty('parent_span_id');
+
+    expect(navigationTraceHeader).toEqual({
+      environment: 'production',
+      public_key: 'public',
+      sample_rate: '1',
+      sampled: 'true',
+      trace_id: navigationTraceContext?.trace_id,
+    });
 
     expect(pageloadTraceContext?.trace_id).not.toEqual(navigationTraceContext?.trace_id);
   },
@@ -52,7 +80,12 @@ sentryTest('error after <meta> tag pageload has pageload traceId', async ({ getL
 
   const url = await getLocalTestPath({ testDir: __dirname });
 
-  const pageloadEvent = await getFirstSentryEnvelopeRequest<Event>(page, url);
+  const [pageloadEvent, pageloadTraceHeader] = await getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+    page,
+    url,
+    eventAndTraceHeaderRequestParser,
+  );
+
   expect(pageloadEvent.contexts?.trace).toMatchObject({
     op: 'pageload',
     trace_id: META_TAG_TRACE_ID,
@@ -60,14 +93,38 @@ sentryTest('error after <meta> tag pageload has pageload traceId', async ({ getL
     span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
   });
 
-  const errorEventPromise = getFirstSentryEnvelopeRequest<Event>(page);
-  await page.locator('#errorBtn').click();
-  const errorEvent = await errorEventPromise;
+  expect(pageloadTraceHeader).toEqual({
+    environment: 'prod',
+    release: '1.0.0',
+    sample_rate: '0.2',
+    sampled: 'true',
+    transaction: 'my-transaction',
+    public_key: 'public',
+    trace_id: META_TAG_TRACE_ID,
+  });
 
-  expect(errorEvent.contexts?.trace).toMatchObject({
+  const errorEventPromise = getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+    page,
+    undefined,
+    eventAndTraceHeaderRequestParser,
+  );
+  await page.locator('#errorBtn').click();
+  const [errorEvent, errorTraceHeader] = await errorEventPromise;
+
+  expect(errorEvent.contexts?.trace).toEqual({
     trace_id: META_TAG_TRACE_ID,
     parent_span_id: META_TAG_PARENT_SPAN_ID,
     span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
+  });
+
+  expect(errorTraceHeader).toEqual({
+    environment: 'prod',
+    release: '1.0.0',
+    sample_rate: '0.2',
+    sampled: 'true',
+    transaction: 'my-transaction',
+    public_key: 'public',
+    trace_id: META_TAG_TRACE_ID,
   });
 });
 
@@ -78,13 +135,20 @@ sentryTest('error during <meta> tag pageload has pageload traceId', async ({ get
 
   const url = await getLocalTestPath({ testDir: __dirname });
 
-  const envelopeRequestsPromise = getMultipleSentryEnvelopeRequests<Event>(page, 2);
+  const envelopeRequestsPromise = getMultipleSentryEnvelopeRequests<EventAndTraceHeader>(
+    page,
+    2,
+    undefined,
+    eventAndTraceHeaderRequestParser,
+  );
   await page.goto(url);
   await page.locator('#errorBtn').click();
-  const events = await envelopeRequestsPromise;
+  const envelopes = await envelopeRequestsPromise;
 
-  const pageloadEvent = events.find(event => event.type === 'transaction');
-  const errorEvent = events.find(event => !event.type);
+  const [pageloadEvent, pageloadTraceHeader] = envelopes.find(
+    eventAndHeader => eventAndHeader[0].type === 'transaction',
+  )!;
+  const [errorEvent, errorTraceHeader] = envelopes.find(eventAndHeader => !eventAndHeader[0].type)!;
 
   expect(pageloadEvent?.contexts?.trace).toMatchObject({
     op: 'pageload',
@@ -93,10 +157,38 @@ sentryTest('error during <meta> tag pageload has pageload traceId', async ({ get
     span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
   });
 
-  expect(errorEvent?.contexts?.trace).toMatchObject({
+  expect(pageloadTraceHeader).toEqual({
+    environment: 'prod',
+    release: '1.0.0',
+    sample_rate: '0.2',
+    sampled: 'true',
+    transaction: 'my-transaction',
+    public_key: 'public',
+    trace_id: META_TAG_TRACE_ID,
+  });
+
+  expect(errorEvent?.contexts?.trace).toEqual({
+    data: {
+      'sentry.op': 'pageload',
+      'sentry.origin': 'auto.pageload.browser',
+      'sentry.sample_rate': 1,
+      'sentry.source': 'url',
+    },
+    op: 'pageload',
+    origin: 'auto.pageload.browser',
     trace_id: META_TAG_TRACE_ID,
     parent_span_id: META_TAG_PARENT_SPAN_ID,
     span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
+  });
+
+  expect(errorTraceHeader).toEqual({
+    environment: 'prod',
+    release: '1.0.0',
+    sample_rate: '0.2',
+    sampled: 'true',
+    transaction: 'my-transaction',
+    public_key: 'public',
+    trace_id: META_TAG_TRACE_ID,
   });
 });
 
@@ -109,12 +201,26 @@ sentryTest(
 
     const url = await getLocalTestPath({ testDir: __dirname });
 
-    const pageloadEvent = await getFirstSentryEnvelopeRequest<Event>(page, url);
+    const [pageloadEvent, pageloadTraceHeader] = await getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+      page,
+      url,
+      eventAndTraceHeaderRequestParser,
+    );
     expect(pageloadEvent?.contexts?.trace).toMatchObject({
       op: 'pageload',
       trace_id: META_TAG_TRACE_ID,
       parent_span_id: META_TAG_PARENT_SPAN_ID,
       span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
+    });
+
+    expect(pageloadTraceHeader).toEqual({
+      environment: 'prod',
+      release: '1.0.0',
+      sample_rate: '0.2',
+      sampled: 'true',
+      transaction: 'my-transaction',
+      public_key: 'public',
+      trace_id: META_TAG_TRACE_ID,
     });
 
     const requestPromise = page.waitForRequest('http://example.com/*');
@@ -137,17 +243,31 @@ sentryTest(
 
     const url = await getLocalTestPath({ testDir: __dirname });
 
-    const pageloadEventPromise = getFirstSentryEnvelopeRequest<Event>(page);
+    const pageloadEventPromise = getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+      page,
+      undefined,
+      eventAndTraceHeaderRequestParser,
+    );
     const requestPromise = page.waitForRequest('http://example.com/*');
     await page.goto(url);
     await page.locator('#fetchBtn').click();
-    const [pageloadEvent, request] = await Promise.all([pageloadEventPromise, requestPromise]);
+    const [[pageloadEvent, pageloadTraceHeader], request] = await Promise.all([pageloadEventPromise, requestPromise]);
 
     expect(pageloadEvent?.contexts?.trace).toMatchObject({
       op: 'pageload',
       trace_id: META_TAG_TRACE_ID,
       parent_span_id: META_TAG_PARENT_SPAN_ID,
       span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
+    });
+
+    expect(pageloadTraceHeader).toEqual({
+      environment: 'prod',
+      release: '1.0.0',
+      sample_rate: '0.2',
+      sampled: 'true',
+      transaction: 'my-transaction',
+      public_key: 'public',
+      trace_id: META_TAG_TRACE_ID,
     });
 
     const headers = request.headers();
@@ -167,12 +287,25 @@ sentryTest(
 
     const url = await getLocalTestPath({ testDir: __dirname });
 
-    const pageloadEvent = await getFirstSentryEnvelopeRequest<Event>(page, url);
+    const [pageloadEvent, pageloadTraceHeader] = await getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+      page,
+      url,
+      eventAndTraceHeaderRequestParser,
+    );
     expect(pageloadEvent?.contexts?.trace).toMatchObject({
       op: 'pageload',
       trace_id: META_TAG_TRACE_ID,
       parent_span_id: META_TAG_PARENT_SPAN_ID,
       span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
+    });
+    expect(pageloadTraceHeader).toEqual({
+      environment: 'prod',
+      release: '1.0.0',
+      sample_rate: '0.2',
+      sampled: 'true',
+      transaction: 'my-transaction',
+      public_key: 'public',
+      trace_id: META_TAG_TRACE_ID,
     });
 
     const requestPromise = page.waitForRequest('http://example.com/*');
@@ -195,17 +328,31 @@ sentryTest(
 
     const url = await getLocalTestPath({ testDir: __dirname });
 
-    const pageloadEventPromise = getFirstSentryEnvelopeRequest<Event>(page);
+    const pageloadEventPromise = getFirstSentryEnvelopeRequest<EventAndTraceHeader>(
+      page,
+      undefined,
+      eventAndTraceHeaderRequestParser,
+    );
     const requestPromise = page.waitForRequest('http://example.com/*');
     await page.goto(url);
     await page.locator('#xhrBtn').click();
-    const [pageloadEvent, request] = await Promise.all([pageloadEventPromise, requestPromise]);
+    const [[pageloadEvent, pageloadTraceHeader], request] = await Promise.all([pageloadEventPromise, requestPromise]);
 
     expect(pageloadEvent?.contexts?.trace).toMatchObject({
       op: 'pageload',
       trace_id: META_TAG_TRACE_ID,
       parent_span_id: META_TAG_PARENT_SPAN_ID,
       span_id: expect.stringMatching(/^[0-9a-f]{16}$/),
+    });
+
+    expect(pageloadTraceHeader).toEqual({
+      environment: 'prod',
+      release: '1.0.0',
+      sample_rate: '0.2',
+      sampled: 'true',
+      transaction: 'my-transaction',
+      public_key: 'public',
+      trace_id: META_TAG_TRACE_ID,
     });
 
     const headers = request.headers();