+++ dev-packages/e2e-tests/test-applications/nextjs-14/app/request-instrumentation/page.tsx
@@ -3,7 +3,7 @@ import http from 'http';
 export const dynamic = 'force-dynamic';
 
 export default async function Page() {
-  await fetch('http://example.com/', { cache: 'no-cache' });
+  await fetch('http://example.com/', { cache: 'no-cache' }).then(res => res.text());
   await new Promise<void>(resolve => {
     http.get('http://example.com/', res => {
       res.on('data', () => {