+++ packages/aws-serverless/src/index.ts
@@ -77,6 +77,8 @@ export {
   setupExpressErrorHandler,
   koaIntegration,
   setupKoaErrorHandler,
+  connectIntegration,
+  setupConnectErrorHandler,
   fastifyIntegration,
   graphqlIntegration,
   mongoIntegration,
@@ -94,6 +96,7 @@ export {
   spanToJSON,
   spanToTraceHeader,
   trpcMiddleware,
+  addOpenTelemetryInstrumentation,
 } from '@sentry/node';
 
 export {