+++ packages/browser/test/integration/suites/shell.js
@@ -22,10 +22,8 @@ function runVariant(variant) {
     /**
      * The test runner will replace each of these placeholders with the contents of the corresponding file.
      */
-    {{ suites/onerror.js }} // biome-ignore format: No trailing commas
     {{ suites/onunhandledrejection.js }} // biome-ignore format: No trailing commas
     {{ suites/builtins.js }} // biome-ignore format: No trailing commas
-    {{ suites/loader.js }} // biome-ignore format: No trailing commas
   });
 }
 
@@ -34,5 +32,3 @@ for (var idx in variants) {
     runVariant(variants[idx]);
   })();
 }
-
-{{ suites/loader-specific.js }} // biome-ignore format: No trailing commas